de-pluralising table names
