+++ docs/tables/oci_certificate_ca_bundle.md
@@ -1,4 +1,4 @@
-# Table: oci_certificates_ca_bundle
+# Table: oci_certificate_ca_bundle
 
 The contents of the certificate authority, properties of the certificate authority (and certificate authority version), and user-provided contextual metadata.
 
@@ -12,7 +12,7 @@ select
     name,
     ca_bundle_pem
 from
-oci_certificates_ca_bundle;
+oci_certificate_ca_bundle;
 ```
 ### Get all Certificate Authority bundles
 
@@ -22,7 +22,7 @@ select
     cca.name,
     cca.ca_bundle_pem
 from
-    oci_certificates_ca_bundle cca
-    inner join oci_certificates_management_ca_bundle cmca
+    oci_certificate_ca_bundle cca
+    inner join oci_certificate_management_ca_bundle cmca
     on cca.ca_bundle_id = cmca.id;
 ```
