+++ docs/tables/oci_certificate_management_ca_bundle.md
@@ -1,4 +1,4 @@
-# Table: oci_certificates_management_ca_bundle
+# Table: oci_certificate_management_ca_bundle
 
 Information and metadata for the certificate authority.
 
@@ -14,5 +14,5 @@ select
     lifecycle_details,
     lifecycle_state as state
 from
-oci_certificates_management_ca_bundle;
+oci_certificate_management_ca_bundle;
 ```
