+++ docs/tables/oci_certificate_management_certificate_version.md
@@ -1,4 +1,4 @@
-# Table: oci_certificates_management_certificate_version
+# Table: oci_certificate_management_certificate_version
 
 The details of the certificate version
 
@@ -19,7 +19,7 @@ select
     validity,
     revocation_status
 from
-oci_certificates_management_certificate_version;
+oci_certificate_management_certificate_version;
 ```
 
 ### Get all certificate versions
@@ -36,7 +36,7 @@ select
     cmcv.validity,
     cmcv.revocation_status
 from
-oci_certificates_management_certificate_version cmcv
-inner join oci_certificates_management_certificate cmc
-on cmcv.certificate_id = cmc.id
+oci_certificate_management_certificate_version cmcv
+inner join oci_certificate_management_certificate cmc
+on cmcv.certificate_id = cmc.id;
 ```
