+++ oci/plugin.go
@@ -49,15 +49,15 @@ func Plugin(ctx context.Context) *plugin.Plugin {
 			"oci_bastion_session":                                          tableBastionSession(ctx),
 			"oci_budget_alert_rule":                                        tableBudgetAlertRule(ctx),
 			"oci_budget_budget":                                            tableBudget(ctx),
-			"oci_certificates_ca_bundle":                                   tableCertificatesCaBundle(ctx),
-			"oci_certificates_certificate_authority_bundle":                tableCertificatesCertificateAuthorityBundle(ctx),
-			"oci_certificates_certificate_bundle":                          tableCertificatesCertificateBundle(ctx),
-			"oci_certificates_management_association":                      tableCertificatesManagementAssociation(ctx),
-			"oci_certificates_management_ca_bundle":                        tableCertificatesManagementCaBundle(ctx),
-			"oci_certificates_management_certificate":                      tableCertificatesManagementCertificate(ctx),
-			"oci_certificates_management_certificate_authority":            tableCertificatesManagementCertificateAuthority(ctx),
-			"oci_certificates_management_certificate_authority_version":    tableCertificatesManagementCertificateAuthorityVersion(ctx),
-			"oci_certificates_management_certificate_version":              tableCertificatesManagementCertificateVersion(ctx),
+			"oci_certificate_ca_bundle":                                    tableCertificatesCaBundle(ctx),
+			"oci_certificate_certificate_authority_bundle":                 tableCertificatesCertificateAuthorityBundle(ctx),
+			"oci_certificate_certificate_bundle":                           tableCertificatesCertificateBundle(ctx),
+			"oci_certificate_management_association":                       tableCertificatesManagementAssociation(ctx),
+			"oci_certificate_management_ca_bundle":                         tableCertificatesManagementCaBundle(ctx),
+			"oci_certificate_management_certificate":                       tableCertificatesManagementCertificate(ctx),
+			"oci_certificate_management_certificate_authority":             tableCertificatesManagementCertificateAuthority(ctx),
+			"oci_certificate_management_certificate_authority_version":     tableCertificatesManagementCertificateAuthorityVersion(ctx),
+			"oci_certificate_management_certificate_version":               tableCertificatesManagementCertificateVersion(ctx),
 			"oci_cloud_guard_configuration":                                tableCloudGuardConfiguration(ctx),
 			"oci_cloud_guard_detector_recipe":                              tableCloudGuardDetectorRecipe(ctx),
 			"oci_cloud_guard_managed_list":                                 tableCloudGuardManagedList(ctx),
