+++ oci/table_oci_certificate_management_certificate_authority_version.go
@@ -11,15 +11,15 @@ import (
 )
 
 // TABLE DEFINITION
-func tableCertificatesManagementCertificateVersion(_ context.Context) *plugin.Table {
+func tableCertificatesManagementCertificateAuthorityVersion(_ context.Context) *plugin.Table {
 	return &plugin.Table{
-		Name:             "oci_certificates_management_certificate_version",
-		Description:      "OCI Certificate Version",
+		Name:             "oci_certificate_management_certificate_authority_version",
+		Description:      "OCI Certificate Authority Version",
 		DefaultTransform: transform.FromCamel(),
 		Get: &plugin.GetConfig{
 			KeyColumns: []*plugin.KeyColumn{
 				{
-					Name:    "certificate_id",
+					Name:    "certificate_authority_id",
 					Require: plugin.Required,
 				},
 				{
@@ -27,13 +27,13 @@ func tableCertificatesManagementCertificateVersion(_ context.Context) *plugin.Ta
 					Require: plugin.Required,
 				},
 			},
-			Hydrate: getCertificatesManagementCertificateVersion,
+			Hydrate: getCertificatesManagementCertificateAuthorityVersion,
 		},
 		List: &plugin.ListConfig{
-			Hydrate: listCertificatesManagementCertificateVersions,
+			Hydrate: listCertificatesManagementCertificateAuthorityVersions,
 			KeyColumns: []*plugin.KeyColumn{
 				{
-					Name:    "certificate_id",
+					Name:    "certificate_authority_id",
 					Require: plugin.Required,
 				},
 				{
@@ -45,18 +45,18 @@ func tableCertificatesManagementCertificateVersion(_ context.Context) *plugin.Ta
 		GetMatrixItemFunc: BuildRegionList,
 		Columns: []*plugin.Column{
 			{
-				Name:        "certificate_id",
-				Description: "The OCID of the certificate.",
+				Name:        "certificate_authority_id",
+				Description: "The OCID of the CA.",
 				Type:        proto.ColumnType_STRING,
 			},
 			{
 				Name:        "version_number",
-				Description: "The version number of the certificate.",
+				Description: "The version number of this CA.",
 				Type:        proto.ColumnType_INT,
 			},
 			{
 				Name:        "stages",
-				Description: "A list of stages of this entity.",
+				Description: "A list of rotation states for this CA version.",
 				Type:        proto.ColumnType_JSON,
 			},
 			{
@@ -66,38 +66,39 @@ func tableCertificatesManagementCertificateVersion(_ context.Context) *plugin.Ta
 			},
 			{
 				Name:        "issuer_ca_version_number",
-				Description: "The version number of the issuing certificate authority (CA).",
+				Description: "The version number of the issuing CA.",
 				Type:        proto.ColumnType_INT,
 			},
 			{
 				Name:        "version_name",
-				Description: "The name of the certificate version. When the value is not null, a name is unique across versions of a given certificate.",
+				Description: "The name of the CA version. When the value is not null, a name is unique across versions for a given CA.",
 				Type:        proto.ColumnType_STRING,
 			},
 			{
 				Name:        "subject_alternative_names",
-				Description: "A list of subject alternative names.",
+				Description: "A list of subject alternative names. A subject alternative name specifies the domain names, including subdomains, and IP addresses covered by the certificates issued by this CA.",
 				Type:        proto.ColumnType_JSON,
+				Hydrate:     getCertificatesManagementCertificateAuthorityVersion,
 			},
 			{
 				Name:        "time_of_deletion",
-				Description: "An optional property indicating when to delete the certificate version.",
+				Description: "An optional property indicating when to delete the CA version.",
 				Type:        proto.ColumnType_TIMESTAMP,
 				Transform:   transform.FromField("TimeOfDeletion.Time"),
 			},
 			{
 				Name:        "validity",
-				Description: "Certificate validity details.",
+				Description: "Certificate Authority validity details.",
 				Type:        proto.ColumnType_JSON,
 			},
 			{
 				Name:        "revocation_status",
-				Description: "Revocation details for the certificate.",
+				Description: "Revocation details for the CA.",
 				Type:        proto.ColumnType_JSON,
 			},
 			{
 				Name:        "time_created",
-				Description: "Time that the Certificate Version was created.",
+				Description: "Time that the Certificate Authority Version was created.",
 				Type:        proto.ColumnType_TIMESTAMP,
 				Transform:   transform.FromField("TimeCreated.Time"),
 			},
@@ -122,10 +123,10 @@ func tableCertificatesManagementCertificateVersion(_ context.Context) *plugin.Ta
 }
 
 // LIST FUNCTION
-func listCertificatesManagementCertificateVersions(ctx context.Context, d *plugin.QueryData, _ *plugin.HydrateData) (interface{}, error) {
+func listCertificatesManagementCertificateAuthorityVersions(ctx context.Context, d *plugin.QueryData, _ *plugin.HydrateData) (interface{}, error) {
 	logger := plugin.Logger(ctx)
 	region := plugin.GetMatrixItem(ctx)[matrixKeyRegion].(string)
-	logger.Debug("listCertificatesManagementCertificateVersions", "OCI_REGION", region)
+	logger.Debug("listCertificatesManagementCertificateAuthorityVersions", "OCI_REGION", region)
 
 	equalQuals := d.KeyColumnQuals
 	// Create Session
@@ -135,7 +136,7 @@ func listCertificatesManagementCertificateVersions(ctx context.Context, d *plugi
 	}
 
 	//Build request parameters
-	request := buildListCertificatesManagementCertificateVersionFilters(equalQuals)
+	request := buildListCertificatesManagementCertificateAuthorityVersionFilters(equalQuals)
 	request.Limit = types.Int(20)
 	request.RequestMetadata = common.RequestMetadata{
 		RetryPolicy: getDefaultRetryPolicy(d.Connection),
@@ -150,7 +151,7 @@ func listCertificatesManagementCertificateVersions(ctx context.Context, d *plugi
 
 	pagesLeft := true
 	for pagesLeft {
-		response, err := session.CertificatesManagementClient.ListCertificateVersions(ctx, request)
+		response, err := session.CertificatesManagementClient.ListCertificateAuthorityVersions(ctx, request)
 		if err != nil {
 			return nil, err
 		}
@@ -173,36 +174,36 @@ func listCertificatesManagementCertificateVersions(ctx context.Context, d *plugi
 }
 
 // HYDRATE FUNCTION
-func getCertificatesManagementCertificateVersion(ctx context.Context, d *plugin.QueryData, h *plugin.HydrateData) (interface{}, error) {
+func getCertificatesManagementCertificateAuthorityVersion(ctx context.Context, d *plugin.QueryData, h *plugin.HydrateData) (interface{}, error) {
 	logger := plugin.Logger(ctx)
 	region := plugin.GetMatrixItem(ctx)[matrixKeyRegion].(string)
-	logger.Debug("getCertificatesManagementCertificateVersion", "OCI_REGION", region)
+	logger.Debug("getCertificatesManagementCertificateAuthorityVersion", "OCI_REGION", region)
 
-	request := buildGetCertificatesManagementCertificateVersionFilters(d.KeyColumnQuals, h)
+	request := buildGetCertificatesManagementCertificateAuthorityVersionFilters(d.KeyColumnQuals, h)
 
 	// Create Session
 	session, err := certificatesManagementService(ctx, d, region)
 	if err != nil {
-		logger.Error("getCertificatesManagementCertificateVersion", "error_CertificatesManagementService", err)
+		logger.Error("getCertificatesManagementCertificateAuthorityVersion", "error_CertificatesManagementService", err)
 		return nil, err
 	}
 	request.RequestMetadata = common.RequestMetadata{
 		RetryPolicy: getDefaultRetryPolicy(d.Connection),
 	}
 
-	response, err := session.CertificatesManagementClient.GetCertificateVersion(ctx, request)
+	response, err := session.CertificatesManagementClient.GetCertificateAuthorityVersion(ctx, request)
 	if err != nil {
 		return nil, err
 	}
-	return response.CertificateVersion, nil
+	return response.CertificateAuthorityVersion, nil
 }
 
 // Build additional list filters
-func buildListCertificatesManagementCertificateVersionFilters(equalQuals plugin.KeyColumnEqualsQualMap) certificatesmanagement.ListCertificateVersionsRequest {
-	request := certificatesmanagement.ListCertificateVersionsRequest{}
+func buildListCertificatesManagementCertificateAuthorityVersionFilters(equalQuals plugin.KeyColumnEqualsQualMap) certificatesmanagement.ListCertificateAuthorityVersionsRequest {
+	request := certificatesmanagement.ListCertificateAuthorityVersionsRequest{}
 
-	if equalQuals["certificate_id"] != nil {
-		request.CertificateId = types.String(equalQuals["certificate_id"].GetStringValue())
+	if equalQuals["certificate_authority_id"] != nil {
+		request.CertificateAuthorityId = types.String(equalQuals["certificate_authority_id"].GetStringValue())
 	}
 
 	if equalQuals["version_number"] != nil {
@@ -213,15 +214,15 @@ func buildListCertificatesManagementCertificateVersionFilters(equalQuals plugin.
 }
 
 // Build additional filters
-func buildGetCertificatesManagementCertificateVersionFilters(equalQuals plugin.KeyColumnEqualsQualMap, h *plugin.HydrateData) certificatesmanagement.GetCertificateVersionRequest {
-	request := certificatesmanagement.GetCertificateVersionRequest{}
+func buildGetCertificatesManagementCertificateAuthorityVersionFilters(equalQuals plugin.KeyColumnEqualsQualMap, h *plugin.HydrateData) certificatesmanagement.GetCertificateAuthorityVersionRequest {
+	request := certificatesmanagement.GetCertificateAuthorityVersionRequest{}
 
 	if h.Item != nil {
-		request.CertificateId = h.Item.(certificatesmanagement.CertificateVersionSummary).CertificateId
-		request.CertificateVersionNumber = h.Item.(certificatesmanagement.CertificateVersionSummary).VersionNumber
+		request.CertificateAuthorityId = h.Item.(certificatesmanagement.CertificateAuthorityVersionSummary).CertificateAuthorityId
+		request.CertificateAuthorityVersionNumber = h.Item.(certificatesmanagement.CertificateAuthorityVersionSummary).VersionNumber
 	} else {
-		request.CertificateId = types.String(equalQuals["certificate_id"].GetStringValue())
-		request.CertificateVersionNumber = types.Int64(equalQuals["version_number"].GetInt64Value())
+		request.CertificateAuthorityId = types.String(equalQuals["certificate_authority_id"].GetStringValue())
+		request.CertificateAuthorityVersionNumber = types.Int64(equalQuals["version_number"].GetInt64Value())
 	}
 
 	return request
